+++ cmd/goto/main.go
@@ -1,0 +1,28 @@
+// Command goto opens file on specific line
+package main
+
+import (
+	"bufio"
+	"fmt"
+	"os"
+	"os/exec"
+	"strings"
+)
+
+func main() {
+	fi, err := os.Stdin.Stat()
+	if err != nil || fi.Mode()&os.ModeNamedPipe == 0 {
+		os.Exit(0)
+	}
+	in := bufio.NewScanner(os.Stdin)
+	for in.Scan() {
+		line := in.Text()
+		part := strings.SplitN(line, " ", 5)
+		if len(part) < 5 {
+			fmt.Fprintln(os.Stderr, line)
+			fmt.Fprintln(os.Stderr, "line should be: FILE FROM TO LINE ...")
+			continue
+		}
+		exec.Command("emacsclient", "-n", "+"+part[3], part[0]).Run()
+	}
+}